Adds new option: -x, --export
The previous "-d" for decrypting with print=true now prints
the decrypted contents so you can populate the `.env` from the
encrypted `.env.enc`.
The new "-x" will print on console the decrypted contents as "export"
statements so you can use this in bash shell scripts to populate your environment
with the `eval` statement.